requirements.txt
transformers>=4.44.0
peft>=0.11.0
accelerate>=1.0.0
datasets>=2.20.0
pillow>=10.4.0
opencv-python-headless>=4.10.0
jsonschema>=4.22.0
pymupdf>=1.24.9
pytesseract>=0.3.13
bitsandbytes
torch>=2.1.0
